+++ compose.yml
@@ -11,7 +11,6 @@ x-service-template: &service-template
     - node_modules_ghost_core:/home/ghost/ghost/core/node_modules:delegated
     - node_modules_ghost_custom-fonts:/home/ghost/ghost/custom-fonts/node_modules:delegated
     - node_modules_ghost_email-addresses:/home/ghost/ghost/email-addresses/node_modules:delegated
-    - node_modules_ghost_email-service:/home/ghost/ghost/email-service/node_modules:delegated
     - node_modules_ghost_i18n:/home/ghost/ghost/i18n/node_modules:delegated
     - node_modules_apps_admin-x-activitypub:/home/ghost/apps/admin-x-activitypub/node_modules:delegated
     - node_modules_apps_admin-x-design-system:/home/ghost/apps/admin-x-design-system/node_modules:delegated
@@ -143,7 +142,6 @@ volumes:
   node_modules_ghost_core: {}
   node_modules_ghost_custom-fonts: {}
   node_modules_ghost_email-addresses: {}
-  node_modules_ghost_email-service: {}
   node_modules_ghost_i18n: {}
   node_modules_apps_admin-x-activitypub: {}
   node_modules_apps_admin-x-design-system: {}
