+++ ghost/core/core/server/services/email-service/EmailEventStorage.js
@@ -67,7 +67,7 @@ class EmailEventStorage {
     /**
      * @private
      * @param {'temporary'|'permanent'} severity
-     * @param {import('@tryghost/email-service').EmailTemporaryBouncedEvent|import('@tryghost/email-service').EmailBouncedEvent} event
+     * @param {import('./events/EmailTemporaryBouncedEvent')|import('./events/EmailBouncedEvent')} event
      * @param {{transacting?: any}} options
      * @returns
      */
