+++ ghost/core/core/server/services/email-service/EmailSegmenter.js
