+++ ghost/core/core/server/services/email-service/events/EmailDeliveredEvent.js
