+++ ghost/core/core/server/services/email-service/events/EmailOpenedEvent.js
