+++ ghost/core/core/server/services/email-service/helpers/register-helpers.js
