+++ ghost/core/package.json
@@ -43,7 +43,7 @@
     "test:browser:setup": "npx playwright install",
     "test:ci:e2e": "c8 -c ./.c8rc.e2e.json -o coverage-e2e yarn test:e2e -b",
     "test:ci:regression": "yarn test:regression -b",
-    "test:ci:integration": "c8 -c ./.c8rc.e2e.json -o coverage-integration --lines 52 --functions 47 --branches 74 --statements 52 yarn test:integration -b",
+    "test:ci:integration": "c8 -c ./.c8rc.e2e.json -o coverage-integration --lines 52 --functions 47 --branches 73 --statements 52 yarn test:integration -b",
     "test:unit:slow": "yarn test:unit --reporter=mocha-slow-test-reporter",
     "test:int:slow": "yarn test:integration --reporter=mocha-slow-test-reporter",
     "test:e2e:slow": "yarn test:e2e --reporter=mocha-slow-test-reporter",
@@ -79,7 +79,6 @@
     "@tryghost/domain-events": "1.0.0",
     "@tryghost/email-addresses": "0.0.0",
     "@tryghost/email-mock-receiver": "0.3.10",
-    "@tryghost/email-service": "0.0.0",
     "@tryghost/errors": "1.3.7",
     "@tryghost/helpers": "1.1.93",
     "@tryghost/html-to-plaintext": "1.0.2",
@@ -238,6 +237,7 @@
     "expect": "29.7.0",
     "find-root": "1.1.0",
     "form-data": "4.0.2",
+    "html-validate": "8.29.0",
     "inquirer": "8.2.6",
     "jwk-to-pem": "2.0.7",
     "jwks-rsa": "3.2.0",
