+++ ghost/core/test/unit/server/services/email-service/batch-sending-service.test.js
@@ -1,5 +1,5 @@
 const {createModel, createModelClass, createDb, sleep} = require('./utils');
-const BatchSendingService = require('../lib/BatchSendingService');
+const BatchSendingService = require('../../../../../core/server/services/email-service/BatchSendingService');
 const sinon = require('sinon');
 const assert = require('assert/strict');
 const logging = require('@tryghost/logging');
@@ -1640,7 +1640,7 @@ describe('Batch Sending Service', function () {
             assert.equal(result, undefined, 'getDeliveryDeadline should return undefined if target delivery window is <=0');
         });
 
-        it('returns undefined if the email.created_at is not set', async function () { 
+        it('returns undefined if the email.created_at is not set', async function () {
             const email = createModel({});
             const service = new BatchSendingService({
                 sendingService: {
