+++ ghost/core/test/unit/server/services/email-service/events/EmailDeliveredEvent.test.js
@@ -1,16 +1,16 @@
 const assert = require('assert/strict');
 const ObjectID = require('bson-objectid').default;
-const EmailOpenedEvent = require('../../lib/events/EmailOpenedEvent');
+const EmailDeliveredEvent = require('../../../../../../core/server/services/email-service/events/EmailDeliveredEvent');
 
-describe('EmailOpenedEvent', function () {
+describe('EmailDeliveredEvent', function () {
     it('exports a static create method to create instances', function () {
-        const event = EmailOpenedEvent.create({
+        const event = EmailDeliveredEvent.create({
             email: 'test@test.test',
             memberId: new ObjectID().toHexString(),
             emailId: new ObjectID().toHexString(),
             emailRecipientId: new ObjectID().toHexString(),
             timestamp: new Date()
         });
-        assert(event instanceof EmailOpenedEvent);
+        assert(event instanceof EmailDeliveredEvent);
     });
 });
