+++ ghost/core/test/unit/server/services/email-service/events/EmailOpenedEvent.test.js
@@ -1,16 +1,16 @@
 const assert = require('assert/strict');
 const ObjectID = require('bson-objectid').default;
-const EmailDeliveredEvent = require('../../lib/events/EmailDeliveredEvent');
+const EmailOpenedEvent = require('../../../../../../core/server/services/email-service/events/EmailOpenedEvent');
 
-describe('EmailDeliveredEvent', function () {
+describe('EmailOpenedEvent', function () {
     it('exports a static create method to create instances', function () {
-        const event = EmailDeliveredEvent.create({
+        const event = EmailOpenedEvent.create({
             email: 'test@test.test',
             memberId: new ObjectID().toHexString(),
             emailId: new ObjectID().toHexString(),
             emailRecipientId: new ObjectID().toHexString(),
             timestamp: new Date()
         });
-        assert(event instanceof EmailDeliveredEvent);
+        assert(event instanceof EmailOpenedEvent);
     });
 });
