+++ ghost/core/test/unit/server/services/email-service/events/EmailUnsubscribedEvent.test.js
@@ -1,6 +1,6 @@
 const assert = require('assert/strict');
 const ObjectID = require('bson-objectid').default;
-const EmailUnsubscribedEvent = require('../../lib/events/EmailUnsubscribedEvent');
+const EmailUnsubscribedEvent = require('../../../../../../core/server/services/email-service/events/EmailUnsubscribedEvent');
 
 describe('EmailUnsubscribedEvent', function () {
     it('exports a static create method to create instances', function () {
