+++ ghost/core/test/unit/server/services/members-events/last-seen-at-updater.test.js
@@ -5,7 +5,7 @@ const LastSeenAtUpdater = require('../../../../../core/server/services/members-e
 const DomainEvents = require('@tryghost/domain-events');
 const {MemberPageViewEvent, MemberCommentEvent, MemberSubscribeEvent, MemberLinkClickEvent} = require('../../../../../core/shared/events');
 const moment = require('moment');
-const {EmailOpenedEvent} = require('@tryghost/email-service');
+const EmailOpenedEvent = require('../../../../../core/server/services/email-service/events/EmailOpenedEvent');
 const EventEmitter = require('events');
 const logging = require('@tryghost/logging');
